Moving Swift 2 tests to a single file
