+++ Tests/SwiftLintFrameworkTests/AttributesRuleTests.swift
@@ -64,18 +64,6 @@ class AttributesRuleTests: XCTestCase {
         verifyRule(alwaysOnNewLineDescription,
                    ruleConfiguration: ["always_on_line_above": ["@objc"]])
     }
-
-    func testSwift2Attributes() {
-        let description = RuleDescription(
-            identifier: AttributesRule.description.identifier,
-            name: AttributesRule.description.name,
-            description: AttributesRule.description.description,
-            nonTriggeringExamples: AttributesRuleExamples.swift2NonTriggeringExamples(),
-            triggeringExamples: AttributesRuleExamples.swift2TriggeringExamples()
-        )
-
-        verifyRule(description)
-    }
 }
 
 extension AttributesRuleTests {
